move smart query string to util
